refactor: let padding_needed_for work purely on u64
To reduce the amount of type conversions, use `u64` for all parameters
and the return type of `padding_needed_for`. Also use `u64` for the
`CPIO_ALIGNMENT` constant.

+++ src/lib.rs
@@ -381,9 +381,8 @@ pub fn list_cpio_content<W: Write>(
 
 /// Returns the amount of padding needed after `offset` to ensure that the
 /// following address will be aligned to `alignment`.
-fn padding_needed_for(offset: u64, alignment: u32) -> u32 {
-    // The rem operation is expected smaller than the right-hand side
-    let misalignment = (offset % u64::from(alignment)) as u32;
+fn padding_needed_for(offset: u64, alignment: u64) -> u64 {
+    let misalignment = offset % alignment;
     if misalignment == 0 {
         return 0;
     }
